requirements.txt
streamlit
plotly
fastapi
uvicorn
sentence_transformers
umap-learn
hdbscan
bs4
aiocache
lxml
trafilatura